requirements.txt
torch>=1.10.0
torchvision>=0.11.0
mmcv-full>=1.5.0
numpy>=1.22.4
opencv-python>=4.5.0
matplotlib>=3.5.0
tqdm>=4.64.0
scikit-learn>=1.1.0
Pillow>=9.0.0
albumentations>=1.3.0
tensorboard>=2.9.0 